--- src/test/java/space/space_spring/domain/authorization/jwt/model/TokenValidatorTest.java
@@ -1,22 +1,0 @@
-package space.space_spring.domain.authorization.jwt.model;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class TokenValidatorTest {
-
-    @Test
-    @DisplayName("")
-    void test() throws Exception {
-        //given
-
-
-        //when
-
-        //then
-    }
-
-
-}
